% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_community_data.R
\name{ob_community_data}
\alias{ob_community_data}
\title{Access onebenthic community data from API 1}
\usage{
ob_community_data(year, sieve_size)
}
\arguments{
\item{year}{The desired survey year}

\item{sieve_size}{The desired sieve size}
}
\value{
The requested output from onebenthic API 1 as \code{\link{str}}
}
\description{
Access onebenthic community data from API 1
}
\examples{
ob_community_data(year = 2000, sieve_size = 1)

}
