% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_faunalcluster_data.R
\name{ob_faunalcluster_data}
\alias{ob_faunalcluster_data}
\title{Access onebenthic faunal cluster ID tool data from API 5}
\usage{
ob_faunalcluster_data(survey_name)
}
\arguments{
\item{survey_name}{The desired survey name}
}
\value{
The requested output from onebenthic API 5 as \code{\link{str}}
}
\description{
Access onebenthic faunal cluster ID tool data from API 5
}
\examples{
ob_faunalcluster_data(survey_name = 'South Coast Regional Seabed Monitoring Programme 2022')

}
