% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_macrofaunal_data.R
\name{ob_macrofaunal_data}
\alias{ob_macrofaunal_data}
\title{Access onebenthic MCZ macrofaunal data from API 4}
\usage{
ob_macrofaunal_data(year)
}
\arguments{
\item{year}{The desired survey year}
}
\value{
The requested output from onebenthic API 4 as \code{\link{str}}
}
\description{
Access onebenthic MCZ macrofaunal data from API 4
}
\examples{
ob_macrofaunal_data(year = 2014)

}
