% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_mtest_data.R
\name{ob_mtest_data}
\alias{ob_mtest_data}
\title{Access onebenthic M-test tool data from API 6}
\usage{
ob_mtest_data(survey_name)
}
\arguments{
\item{survey_name}{The desired survey name}
}
\value{
The requested output from onebenthic API 6 as \code{\link{str}}
}
\description{
Access onebenthic M-test tool data from API 6
}
\examples{
ob_mtest_data(survey_name = 'Eastern English Channel Regional Seabed Monitoring Programme 2020')

}
