% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_nonnative_data.R
\name{ob_nonnative_data}
\alias{ob_nonnative_data}
\title{Access onebenthic non-native species data from API 7}
\usage{
ob_nonnative_data(year)
}
\arguments{
\item{year}{The desired survey year}
}
\value{
The requested output from onebenthic API 7 as \code{\link{str}}
}
\description{
Access onebenthic non-native species data from API 7
}
\examples{
ob_nonnative_data(year = 2014)

}
