% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_sediment_data.R
\name{ob_sediment_data}
\alias{ob_sediment_data}
\title{Access onebenthic sediment data from API 3}
\usage{
ob_sediment_data(year)
}
\arguments{
\item{year}{The desired survey year}
}
\value{
The requested output from onebenthic API 3 as \code{\link{str}}
}
\description{
Access onebenthic sediment data from API 3
}
\examples{
ob_sediment_data(year = 2000)

}
