% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ob_taxon_data.R
\name{ob_taxon_data}
\alias{ob_taxon_data}
\title{Access onebenthic taxon data from API 2}
\usage{
ob_taxon_data(aphiaid)
}
\arguments{
\item{aphiaid}{The desired species aphiaid}
}
\value{
The requested output from onebenthic API 2 as \code{\link{str}}
}
\description{
Access onebenthic taxon data from API 2
}
\examples{
ob_taxon_data(aphiaid = 130867)

}
